+++ src/id_generator.h
@@ -3,6 +3,6 @@
 
 #include "stdint.h"
 
-uint32_t get_sessionid();
+uint64_t get_sessionid();
 
 #endif  //! #ifndef ID_GENERATOR_H_
